Accept bigint in pilo_member_count_band so COUNT(*) calls match
COUNT(*) returns bigint in Postgres and the function was declared with
integer, so the implicit-cast resolution failed at runtime with
'function pilo_member_count_band(bigint) does not exist'. Switching
the parameter to bigint is the cleanest fix — every numeric caller
either is a bigint already or coerces freely up to bigint.

+++ supabase/migrations/017_pilo_cohort_populator.sql
@@ -10,7 +10,7 @@
 -- ────────────────────────────────────────────────────────────
 -- Helper: member-count band for a group of N approved+managed members
 -- ────────────────────────────────────────────────────────────
-CREATE OR REPLACE FUNCTION pilo_member_count_band(n integer)
+CREATE OR REPLACE FUNCTION pilo_member_count_band(n bigint)
 RETURNS text LANGUAGE sql IMMUTABLE AS $$
   SELECT CASE
     WHEN n < 10  THEN 'lt_10'
